extract prop suggestion selector
